# server1/swagger.json
{
    "openapi":"3.0.0",
    "info":{
        "title":"API de manpulação do Motor de Vendas",
        "description":"Essa API tem como objetivo a manipulação do Motor de Vendas",
        "version":"1.0.0"
    },
    "servers":[
        {
            "url":"http://localhost:3007",
            "description":"API de teste"
        }
    ],
    "paths":{
...
        "/group":{
            "post":{
                "description":"Cadastrar grupo",
                "summary":"Essa rota é responsável por cadastrar um grupo",
                "tags":["Grupo"],
                "security":[{"bearerAuth":[]}],
                "requestBody":{
                    "content":{
                        "application/json":{
                            "schema":{
                                "type":"object",
                                "$ref":"#/components/schemas/Group",
                                "required":true
                            },
                            "examples":{
                                "grupo":{
                                    "value":{
                                        "description":"Calça"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses":{
                    "400":{
                        "description":"Request needs a token"
                    },
                    "404":{
                        "description":"Token invalid. Its necessary bearer + token"

                    },
                    "401": {
                        "description":"product already exists!"
                    },
                    "403":{
                        "description":"Invalid token"

                    },
                    "200":{
                        "description": "ok"
                        
                    }
                }


            },
            "get":{
                "description":"Buscar todos grupos",
                "summary":"Essa rota é responsável por buscar todos grupos",
                "tags":["Grupo"],
                "security":[{"bearerAuth":[]}],
                "parameters": [],
                "responses": {
                    "200":{
                        "description": "OK",
                        "content": {
                            "application/json":{
                                "schema":{
                                    "type":"array",
                                    "items":{
                                        "$ref": "#/components/schemas/Group"
                                    }
                                    
                                }
                            }
                        }
                    },
                    "400":{
                        "description":"Request needs a token"
                    },
                    "404":{
                        "description":"Token invalid. Its necessary bearer + token"
    
                    },
                    "403":{
                        "description":"Invalid token"
    
                    }
                }


            },
            "put":{
                "description":"Edita um grupo",
                "summary":"Essa rota é responsável por editar um único grupo passando seu id pelo body",
                "tags":["Grupo"],
                "security":[{"bearerAuth":[]}],

                "requestBody":{
                    "content":{
                        "application/json":{
                            "schema":{
                                "type": "object",
                                "properties": {
                                
                                    "description":{
                                        "type": "string",
                                        "example":"grupo1",
                                        "description": "Descrição do grupo"
                                    },
                                    "id":{
                                        "type":"integer",
                                        "example":"1",
                                        "description": "id do grupo"
                                    }

                                },
                                "required": [
                                    "description","id"
                                ]

                            }
                        }
                    }
                },
                "responses":{
                    "200":{
                        "description":"OK"
                    },
                    "400":{
                        "description":"Request needs a token"
                    },
                    "404":{
                        "description":"Token invalid. Its necessary bearer + token"
    
                    },
                    "403":{
                        "description":"Invalid token"
    
                    }
                }

            }
        },
        "/group/{id}":{
            "patch":{
                "description": "Edita o 'status' do grupo para '0' pelo id",
                "summary": "Essa rota é responsável por inatival o produto pelo id",
                "tags": ["Grupo"],
                "security":[{"bearerAuth":[]}],
                "parameters": [{
                        "name":"id",
                        "in":"path",
                        "description": "grupo pelo id",
                        "required": true
                    }
                ],
                "responses": {
                    "400":{
                        "description":"Request needs a token"
                    },
                    "200":{
                        "description": "OK"
                    },
                    "404":{
                        "description":"Token invalid. Its necessary bearer + token"

                    },
                    "403":{
                        "description":"Invalid token"

                    }

                }

            },
            "get":{
                "description": "Buscar um único grupo pelo id",
                "summary": "Essa rota é responsável por buscar um único grupo pelo id",
                "tags": ["Grupo"],
                "security":[{"bearerAuth":[]}],
                "parameters": [{
                        "name":"id",
                        "in":"path",
                        "description": "grupo pelo id",
                        "required": true
                    }
                ],
                "responses": {
                    "200":{
                        "description": "OK",
                        "content": {
                            "application/json":{
                                "schema":{
                                    "type":"array",
                                    "items":{
                                        "$ref": "#/components/schemas/Group"
                                    }
                                    
                                }
                            }
                        }
                    },
                    "400":{
                        "description":"Request needs a token"
                    },
                    "404":{
                        "description":"Token invalid. Its necessary bearer + token"
    
                    },
                    "403":{
                        "description":"Invalid token"
    
                    }
                }

            }
        },
...
        "/collection":{
            "post":{
                "summary":"Cadastro  de coleção",
                "description":"Essa rota é responsável pelo cadastro de uma nova coleção",
                "tags":["Coleção"],
                "security":[{"bearerAuth":[]}],
                "requestBody":{
                    "content":{
                        "application/json":{
                            "schema":{
                                "type": "object",
                                "$ref":"#/components/schemas/Collection"
                                        
                            },
                            "examples":{
                                "subgroup":{
                                    "value":{
                                        "description":"Inverno 2023"     
                                    }
                                 }
                            }
                        }
                    }
                },
                "responses":{
                    "400":{
                        "description":"Request needs a token"
                    },
                    "404":{
                        "description":"Token invalid. Its necessary bearer + token"

                    },
                    "401": {
                        "description":"product already exists!"
                    },
                    "403":{
                        "description":"Invalid token"

                    },
                    "200":{
                        "description": "ok"
                        
                    }
                }
            },
            "get":{
                "description":"Buscar todas coleções",
                "summary":"Essa rota é responsável por buscar todas coleções",
                "tags":["Coleção"],
                "security":[{"bearerAuth":[]}],
                "parameters": [],
                "responses": {
                    "200":{
                        "description": "OK",
                        "content": {
                            "application/json":{
                                "schema":{
                                    "type":"array",
                                    "items":{
                                        "$ref": "#/components/schemas/Collection"
                                    }
                                    
                                }
                            }
                        }
                    },
                    "400":{
                        "description":"Request needs a token"
                    },
                    "404":{
                        "description":"Token invalid. Its necessary bearer + token"
    
                    },
                    "403":{
                        "description":"Invalid token"
    
                    }
                }
            },
            "put":{
                "description":"Edita uma coleção",
                "summary":"Essa rota é responsável por editar uma única coleção passando seu id pelo body",
                "tags":["Coleção"],
                "security":[{"bearerAuth":[]}],
                "requestBody":{
                    "content":{
                        "application/json":{
                            "schema":{
                                "type": "object",
                                "properties": {
                                    "description":{
                                        "type": "string",
                                        "example":"Inverno 2024",
                                        "description": "Descrição da coleção"
                                    },
                                    "id":{
                                        "type":"integer",
                                        "example":1,
                                        "description":"id da coleção"
                                    }
                                },
                                "required": [
                                    "description","id"
                                ]
                            }
                        }
                    }
                },
                "responses":{
                    "200":{
                        "description":"OK"
                    },
                    "400":{
                        "description":"Request needs a token"
                    },
                    "404":{
                        "description":"Token invalid. Its necessary bearer + token"
    
                    },
                    "403":{
                        "description":"Invalid token"
    
                    }
                }
            }
        },
        "/collection/{id}":{
            "get":{
                "description": "Buscar uma única coleção pelo id",
                "summary": "Essa rota é responsável por buscar uma única coleção pelo id",
                "tags": ["Coleção"],
                "security":[{"bearerAuth":[]}],
                "parameters": [{
                        "name":"id",
                        "in":"path",
                        "description": "id da coleção",
                        "required": true
                    }
                ],
                "responses": {
                    "200":{
                        "description": "OK",
                        "content": {
                            "application/json":{
                                "schema":{
                                    "type":"array",
                                    "items":{
                                        "$ref": "#/components/schemas/Collection"
                                    }
                                }
                            }
                        }
                    },
                    "400":{
                        "description":"Request needs a token"
                    },
                    "404":{
                        "description":"Token invalid. Its necessary bearer + token"
    
                    },
                    "403":{
                        "description":"Invalid token"
    
                    }
                }
            },
            "patch":{
                "description": "Edita o 'status' de uma coleção para '0' pelo id",
                "summary": "Essa rota é responsável por inatival a coleção  pelo id",
                "tags": ["Coleção"],
                "security":[{"bearerAuth":[]}],
                "parameters": [{
                        "name":"id",
                        "in":"path",
                        "description": "id da coleção",
                        "required": true
                    }
                ],
                "responses": {
                    "400":{
                        "description":"Request needs a token"
                    },
                    "200":{
                        "description": "OK"
                    },
                    "404":{
                        "description":"Token invalid. Its necessary bearer + token"

                    },
                    "403":{
                        "description":"Invalid token"

                    }

                }
            }
        },
...
    },
    "components":{
        "schemas":{
...
            "Group":{
                "required":["description"],
                "properties":{
                    "description":{
                        "type":"string",
                        "description":"Descrição do grupo"
                    }
                }  
            },
...
            "Collection":{
                "required":["description"],
                "properties":{
                    "decription":{
                        "type":"string",
                        "description":"Descrição da coleção"
                    }
                }
            },
...
        },
        "securitySchemes":{
            "bearerAuth":{
                "type":"http",
                "scheme":"bearer",
                "bearerFormat":"JWT"
            }
        }
    }
}